refactor server and config files
