+++ public/css/site.styl
@@ -10,3 +10,8 @@ footer
     padding-right 15px
     p
         border-top: 1px solid #d7e0e2
+.navbar-form
+    .form-group
+        margin-left: 10px
+    button 
+        margin-left: 10px
