+++ server/config/routes.js
@@ -1,0 +1,13 @@
+
+
+module.exports = function (app) {
+    app.get('/partials/*', function(req, res){
+        res.render('../../public/app/' + req.params[0])
+    })
+    
+    app.get('*', function(req, res) {
+        res.render('index', {
+            // mongoMessage: mongoMessage
+        });
+    })
+}
